requirements.txt
Django>=4.2,<5.0
requests>=2.31.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
